robot-server/requirements.txt
python-socketio==5.7.1

eventlet
numpy 
Pillow
PiCamera